% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{agent_requirement}
\alias{agent_requirement}
\title{agent_requirement}
\format{An object of class \code{matrix} with 48 rows and 7 columns.}
\usage{
agent_requirement
}
\description{
a 48 x 7 matrix to specifiy number of agents needed at each 0.5 hr each day,
 serve as an example data for demonstrating scheduler.
}
\keyword{datasets}

