% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.r
\name{agentActiveHours}
\alias{agentActiveHours}
\title{agentActiveHours}
\usage{
agentActiveHours(sm, j, i)
}
\arguments{
\item{sm}{schedule module - single value:
schedule module 1: agent start any half hour window, say 8:30AM,
  work 3.5 hour, take 0.5 hour break, and work another 4.5 hour.
schedule module 2: agent start any half hour window, say 8:30AM,
  work 4.0 hour, take 0.5 hour break, and work another 4.0 hour.
schedule module 3: agent start any half hour window, say 8:30AM,
  work 4.5 hour, take 0.5 hour break, and work another 3.5 hour.}

\item{j}{start day j = 0L Sunday, 1L Monday, ..., or 6L Saturday.}

\item{i}{start hour i = 0L 00:00AM, 2L 00:30AM, ..., or 47L 11:30AM.}
}
\description{
a subroutine in scheduler for calculating agent active hour given a starting
 half hour window and day, and the schedule module.
 for example, an agent start at 07:30AM Sunday and schedule module 1 would be
 available at 07:30AM - 11:00AM and 11:30AM - 16:00PM Monday - Friday, so the
 active half-hour indx would be 16-22, 24-32, 64-70, 72-80, 112-118, 120-128,
 160-166, 168-176, 208-214 and 216-224.
}
\note{
note that there is a slight inconsistency that i is 0 - 47 half hour index
 and j is 0 - 6 day index, so the full half hour index produced should then
 be 0-335 but we add 1 and return 1-336 as lp model take row indexed from 1
}
\seealso{
Other scheduler: \code{\link{agentOnTimeFromSchedule}},
  \code{\link{scheduler_solution_constructor}},
  \code{\link{scheduler}}
}

