% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.r
\name{agentOnTimeFromSchedule}
\alias{agentOnTimeFromSchedule}
\title{agentOnTimeFromSchedule}
\usage{
agentOnTimeFromSchedule(ss_lst)
}
\description{
a subroutine in scheduler_solution_constructor to construct ss from s1, s2, s3
}
\seealso{
Other scheduler: \code{\link{agentActiveHours}},
  \code{\link{scheduler_solution_constructor}},
  \code{\link{scheduler}}
}

