% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{compare_array}
\alias{compare_array}
\title{Compare a vector of hierarchical categories.}
\usage{
compare_array(x, set_missing = "Scheduled")
}
\arguments{
\item{x}{a character vector representing names of hierarchical categories.}

\item{set_missing}{a value or set of values to indicate the result should be 
missing.}
}
\value{
a character vector of length one specifying the hierarchical category
  chosen.
}
\description{
Compare a vector of hierarchical categories.
}
\details{
The participants are asked to provide what dates they are available 
  and what date in particular is best for them. We also are scheduling 
  participants all the time. Thus, we have a hierarchical structure for 
  deciding participants: Unavialable/Available < Preference < Scheduled. If 
  we have already scheduled the date, then we need to set that value as 
  missing so we know we can't schedule it.
  
  This is for use when creating a final table of availability since the
  original data is in a 3D array with dates, guests, and availability
  category.
}
\examples{
compare_array(c("Available", "Preference", NA))
compare_array(c("Available", NA, "Scheduled"))
compare_array(c("Unavailable", NA, NA))
}

