% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{count_unique_meetings}
\alias{count_unique_meetings}
\title{Counts per person how many unique people they meet}
\usage{
count_unique_meetings(availability_array, is_host = TRUE)
}
\arguments{
\item{availability_array}{3d array of availability with entries 0, 1, and 2}

\item{is_host}{If TRUE, do for host; otherwise, do guest}
}
\value{
a vector of the number of unique people each person meets
}
\description{
Counts per person how many unique people they meet
}
