% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill gaps}
\usage{
fill_gaps(
  availability_array,
  too_long = 4,
  desired_length = 60,
  slot_length = 15,
  earliest_possible = TRUE,
  host_rooms,
  excluded_hosts = c(),
  included_hosts = c(),
  allow_shorter_meetings = FALSE,
  max_tries = 1000,
  add_gap = FALSE
)
}
\arguments{
\item{availability_array}{3d array of availability with entries 0, 1, and 2}

\item{too_long}{The maximum number of slots a gap should be}

\item{desired_length}{The amount of time to require in a slot}

\item{slot_length}{The amount of time each slot represents}

\item{earliest_possible}{If TRUE, tries to do this meeting as early in the day as it can; if FALSE, as late}

\item{host_rooms}{The vector of host rooms: room is entry, host name is names}

\item{excluded_hosts}{A vector of hosts who shouldn't be scheduled}

\item{included_hosts}{A vector of hosts who should be scheduled. Only include one of these two.}

\item{allow_shorter_meetings}{If TRUE, allow meetings shorter than desired_length}

\item{add_gap}{If TRUE, add a gap between meetings of one slot if possible}
}
\value{
a 3d array of availability with entries 0, 1, and 2
}
\description{
Fill gaps
}
