% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{get_availability}
\alias{get_availability}
\title{Takes in comma separated months and days, returns a vector of POSIX dates.}
\usage{
get_availability(x)
}
\arguments{
\item{x}{a string of comma separated dates.}
}
\value{
a vector of POSIX dates.
}
\description{
Takes in comma separated months and days, returns a vector of POSIX dates.
}
\details{
One of the items in the survey asks the participants to choose the
  date that they are available. This is rendered as a single column with
  comma separated values.
}
\examples{
get_availability(c("Jan 4 2015, Sep 6, March 20 2016"))
}

