% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{get_fills_per_host}
\alias{get_fills_per_host}
\title{Get the maximum number of adjacent slots a guest has filled}
\usage{
get_fills_per_host(availibility_array)
}
\arguments{
\item{availibility_array}{3d array of availability with entries 0, 1, and 2}
}
\value{
The maximum number of adjacent slots a guest has filled
}
\description{
This ignores things like lunch (i.e., if there is a gap before lunch and two after, it will return 3)
}
