% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_last_sunday}
\alias{get_last_sunday}
\title{Return the date for last Sunday}
\usage{
get_last_sunday(now = Sys.Date())
}
\arguments{
\item{now}{Today's date in ymd format. Defaults to the output of \code{Sys.Date()}.}
}
\value{
Last Sunday's date in POSIXct format
}
\description{
Return the date for last Sunday
}
\examples{
get_last_sunday()
get_last_sunday("2014-02-13")
}

