% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{get_max_gap_per_guest}
\alias{get_max_gap_per_guest}
\title{Get the maximum number of adjacent slots a guest is available for}
\usage{
get_max_gap_per_guest(availibility_array, guest_name)
}
\arguments{
\item{availibility_array}{3d array of availability with entries 0, 1, and 2}

\item{guest_name}{The name of the guest}
}
\value{
The maximum number of adjacent slots a guest is available for
}
\description{
This ignores things like lunch (i.e., if there is a gap before lunch and two after, it will return 3)
}
