% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{guest_availability}
\alias{guest_availability}
\title{Gathers guest input data and computes availability array}
\usage{
guest_availability(data, slot_minutes = 15, input_minutes = 30)
}
\arguments{
\item{data}{Input CSV file}

\item{slot_minutes}{How long to make each slot}

\item{input_minutes}{Resolution of the input slots}
}
\value{
an availability array
}
\description{
This takes a CSV with columns for Name, Room, Email, and Times (separated by commas and spaces) and Essential and Optional people to meet with (separated by commas and spaces).This sort of file could come from a Google Sheet, for example, where hosts have filled in information about times in a form with checkboxes. If there are times like 8:30 am, 9 am, etc. and slot_minutes are 15 minutes, it will assume that if someone selects 8:30 they can make both 8:30 and 8:45 slots. If you want more precise slots, you can have slot_minutes less than input_minutes: they said this, say, 30 minute slot was available, which means that both 15 minute slots within could be free.
}
