% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{infohover}
\alias{infoclick}
\alias{infohover}
\title{Display function for html visualization}
\usage{
infohover(x)

infoclick(x)
}
\arguments{
\item{x}{a data frame used to create the ggvis object.}
}
\value{
a string encoding html.
}
\description{
The info tip will display information of the participants.
}

