% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{infotip}
\alias{infotip}
\title{Display function for html visualization}
\usage{
infotip(x)
}
\arguments{
\item{x}{a data frame used to create the ggvis object.}
}
\value{
a string encoding html.
}
\description{
The info tip will display information of the participants.
}

