% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{make_dossier}
\alias{make_dossier}
\alias{make_filename}
\title{Function to create a markdown dossier for each participant}
\usage{
make_dossier(x, df, xlist, wd = ".")

make_filename(x, wd = ".", newdir = "/dossiers/")
}
\arguments{
\item{x}{a name of a participant}

\item{df}{a data frame containing relevant information}

\item{xlist}{a list with one element per participant containing the dates for
availability.}

\item{wd}{the location of the directory that contains the 'dossiers' folder.}
}
\value{
nothing.
}
\description{
Function to create a markdown dossier for each participant
}

