% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{make_my_day}
\alias{make_my_day}
\title{Pretty print month day year}
\usage{
make_my_day(Date)
}
\arguments{
\item{Date}{a date in POSIXct format}
}
\value{
A character string
}
\description{
This will print the month, day, and year for any Date object
}
\examples{
make_my_day(ymd(Sys.Date()))
}

