% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{possible_availability_create}
\alias{possible_availability_create}
\title{Create array of hosts, guests, and possible times}
\usage{
possible_availability_create(host_availability, guest_availability)
}
\arguments{
\item{host_availability}{Availability matrix for hosts}

\item{guest_availability}{Availability matrix for guests}
}
\value{
3D array with dimensions host, guest, times
}
\description{
This will take an availability array of hosts and guests and delete any times that aren't possible for any host-guest pair
}
