% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{save_the_date}
\alias{save_the_date}
\title{Takes in a vector of sundays and returns a logical vector stating which
Sundays in the year the person is available.}
\usage{
save_the_date(x, any_given_sunday)
}
\arguments{
\item{x}{a vector of POSIX dates produced from \code{\link{get_availability}}}

\item{any_given_sunday}{a vector of POSIX dates defining the range of
possible Sundays.}
}
\value{
a vector of logical elements indicating which dates are available for
  a person.
}
\description{
Takes in a vector of sundays and returns a logical vector stating which
Sundays in the year the person is available.
}

