% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule_durations}
\alias{schedule_durations}
\title{Convert schedule to easier information for a person including durations}
\usage{
schedule_durations(simple_schedule, input_minutes = 20)
}
\arguments{
\item{simple_schedule}{A data.frame of times, the person being met, and the room}

\item{input_minutes}{The amount of time each slot represents}
}
\value{
A data.frame of times, the person being met, the room, and the duration
}
\description{
Convert schedule to easier information for a person including durations
}
