% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule_flatten}
\alias{schedule_flatten}
\title{Flatten everyone's schedules to one data.frame}
\usage{
schedule_flatten(availability_array, is_host = TRUE)
}
\arguments{
\item{availability_array}{3d array of availability with entries 0, 1, and 2}

\item{is_host}{If TRUE, do for host; otherwise, do guest}
}
\value{
a data.frame of times and who each person (host or guest) is meeting with at each time
}
\description{
Flatten everyone's schedules to one data.frame
}
