% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.r
\name{schedule_viewer}
\alias{schedule_viewer}
\title{schedule_viewer}
\usage{
schedule_viewer(m, xlab_text = "Half-Hour Window",
  ylab_text = "Number of Agents",
  ggtitle_text = "Number of Agents by Half-Hour Window at Each Day",
  element_text_size = 22L)
}
\arguments{
\item{m}{- a 48 x 7 matrix}
}
\value{
a ggplot object that plot 48 half hour window each day for 7 day
}
\description{
a visualization funtion to show agent requirement or schedules from
 48 x 7 matrix
}
\seealso{
Other scheduler_viewer: \code{\link{schedule_viewer1}},
  \code{\link{schedule_viewer2}},
  \code{\link{schedule_viewer3}}
}

