% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.r
\name{schedule_viewer2}
\alias{schedule_viewer2}
\title{schedule_viewer2}
\usage{
schedule_viewer2(m1, m2, m1_label = "Agent Required",
  m2_label = "Agent Available", legend_title = "Legend",
  xlab_text = "Half-Hour Window", ylab_text = "Number of Agents",
  ggtitle_text = "Number of Agents by Half-Hour Window at Each Day",
  element_text_size = 22L)
}
\arguments{
\item{m1}{a 48 x 7 matrix}

\item{m2}{a 48 x 7 matrix}
}
\value{
a ggplot object that plot 48 half hour window each day for 7 day
}
\description{
a visualization funtion to show agent requirement or schedules from two
 48 x 7 matrices, compare requirement and schedule side by side
}
\seealso{
Other scheduler_viewer: \code{\link{schedule_viewer1}},
  \code{\link{schedule_viewer3}},
  \code{\link{schedule_viewer}}
}

