% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.r
\name{schedule_viewer3}
\alias{schedule_viewer3}
\title{schedule_viewer3}
\usage{
schedule_viewer3(m1, m2, m3, m1_label = "Schedule Module 1",
  m2_label = "Schedule Module 2", m3_label = "Schedule Module 3",
  legend_title = "Legend", xlab_text = "Half-Hour Window",
  ylab_text = "Number of Agents",
  ggtitle_text = "Number of Agents by Half-Hour Window at Each Day",
  element_text_size = 22L)
}
\arguments{
\item{m1, }{m2, m3 - 48 x 7 matrices}
}
\value{
a ggplot object that plot 48 half hour window each day for 7 day
}
\description{
a visualization funtion to show agent requirement or schedules from 3
 48 x 7 matrices, view details of 3 schedule modules side by side
}
\seealso{
Other scheduler_viewer: \code{\link{schedule_viewer1}},
  \code{\link{schedule_viewer2}},
  \code{\link{schedule_viewer}}
}

