% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{scheduled}
\alias{IDS}
\alias{scheduled}
\title{Global Data}
\format{An object of class \code{list} of length 2.}
\usage{
scheduled$get()

IDS$get()
}
\description{
a list stored in an internal environment
}
\keyword{datasets}

