% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleR-package.R
\docType{package}
\name{scheduleR-package}
\alias{scheduleR}
\alias{scheduleR-package}
\title{scheduleR: R Interface to Windows Task Scheduler}
\description{
Provides an R API wrapper around the Windows 'schtasks' command-line tool for creating, querying, modifying, and managing scheduled tasks on Windows systems.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
