% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.r
\name{scheduler_solution_constructor}
\alias{scheduler_solution_constructor}
\title{scheduler_solution_constructor}
\usage{
scheduler_solution_constructor(scheduleOptModel, sm)
}
\description{
a subroutine in scheduler for constructing scheduler solution from lp solution
}
\seealso{
Other scheduler: \code{\link{agentActiveHours}},
  \code{\link{agentOnTimeFromSchedule}},
  \code{\link{scheduler}}
}

