% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{time_difference}
\alias{time_difference}
\title{Compute the minutes between two times in HH:MM format}
\usage{
time_difference(times)
}
\arguments{
\item{times}{The vector the two times}
}
\value{
The difference in minutes
}
\description{
Compute the minutes between two times in HH:MM format
}
