% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{time_interpolate}
\alias{time_interpolate}
\title{More finely divide a time slot}
\usage{
time_interpolate(time, slot_minutes = 15, input_minutes = 30)
}
\arguments{
\item{time}{The string of the time}

\item{slot_minutes}{How long to make each slot}

\item{input_minutes}{Resolution of the input slots}
}
\value{
The hour:minute
}
\description{
If people give the hours available, and you want to split those up into 5 minute segments, this will do that.
}
