% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{times_format}
\alias{times_format}
\title{Converts times}
\usage{
times_format(time, assume_pm = 7)
}
\arguments{
\item{time}{The string of the time}

\item{assume_pm}{The latest hour at which you assume the time means PM if not otherwise specified}
}
\value{
The hour:minute
}
\description{
Handles times like 8, 8 AM, 8 pm, 8:30 am, 8:30 PM and converts them all to the hour and minute, with the hour being a 24 hour clock. If you have a time like "2:00" it looks at the hour; if it's before assume_pm, it assumes you mean 2 pm, not 2 am
}
