% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{availability_fill}
\alias{availability_fill}
\title{Fill in meetings}
\usage{
availability_fill(
  possible_availability,
  guests,
  desired_length = 60,
  slot_length = 15,
  earliest_possible = TRUE
)
}
\arguments{
\item{possible_availability}{3D array with dimensions host, guest, times}

\item{guests}{The data.frame of guest info, including a column of Name of the guest and Desired the hosts to meet}

\item{desired_length}{The amount of time to require in a slot}

\item{slot_length}{The amount of time each slot represents}

\item{earliest_possible}{If TRUE, tries to do this meeting as early in the day as it can; if FALSE, as late}
}
\value{
An array in same format as possible_availability, but with 2 for the assigned slots, 0 for the unavailable slots, and 1 for available but still unfilled.
}
\description{
This will take an availability array and info on guest priorities and fill in the essential meetings. It will make sure the slots are continuous.
}
