% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{scheduled}
\alias{IDS}
\alias{scheduled}
\title{Global Data}
\format{\preformatted{List of 2
 $ get:function (v)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 123 11 123 57 11 57 123 123
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fd23b452788> 
 $ set:function (v)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 124 11 124 34 11 34 124 124
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fd23b452788> 
}}
\usage{
scheduled$get()

IDS$get()
}
\description{
a list stored in an internal environment
}
\keyword{datasets}

