% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{scheduled}
\alias{IDS}
\alias{scheduled}
\title{Global Data}
\format{\preformatted{List of 2
 $ get:function (v)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 154 11 154 57 11 57 154 154
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe1ed65c670> 
 $ set:function (v)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 155 11 155 34 11 34 155 155
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fe1ed65c670> 
}}
\usage{
scheduled$get()

IDS$get()
}
\description{
a list stored in an internal environment
}
\keyword{datasets}

