% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{scheduled}
\alias{IDS}
\alias{scheduled}
\title{Global Data}
\format{\preformatted{List of 2
 $ get:function (v)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 161 11 161 57 11 57 161 161
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb24e1a6a70> 
 $ set:function (v)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 162 11 162 34 11 34 162 162
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7fb24e1a6a70> 
}}
\usage{
scheduled$get()

IDS$get()
}
\description{
a list stored in an internal environment
}
\keyword{datasets}

